backend-fastapi/requirements.txt
fastapi
uvicorn
SQLModel
asyncio-throttle
python-binance
aiohttp
channels
asgiref
sqlalchemy
aiosqlite